Added LayerConstruct custom validation, added IComputationHandler.AsNumber(ndarray, indices) and AsNDArray(number)
Added LayerConstruct custom validation, added
IComputationHandler.AsNumber(ndarray, indices) and AsNDArray(number)

+++ Sigma.Core/Layers/InputLayer.cs
@@ -48,7 +48,7 @@ namespace Sigma.Core.Layers
 			construct.ExternalInputs = new[] { inputAlias };
 			construct.Parameters["external_input_alias"] = inputAlias;
 			construct.Parameters["shape"] = shape;
-			construct.Parameters["size"] = ArrayUtils.Product(shape);
+			construct.Parameters["size"] = (int) ArrayUtils.Product(shape);
 
 			return construct;
 		}
