+++ Sigma.Core/Layers/OutputLayer.cs
@@ -49,7 +49,7 @@ namespace Sigma.Core.Layers
 			construct.ExternalOutputs = new[] { externalOutputAlias };
 			construct.Parameters["external_output_alias"] = externalOutputAlias;
 			construct.Parameters["shape"] = shape;
-			construct.Parameters["size"] = ArrayUtils.Product(shape);
+			construct.Parameters["size"] = (int) ArrayUtils.Product(shape);
 
 			return construct;
 		}
